chore : 기존 인수 테스트 삭제
